delay checking for cli commands until needed
