+++ src/types.ts
@@ -32,3 +32,7 @@ export { CompletionItems, SpellCheckScope, Snippet, Snippets };
 
 // Set up types:
 export type LabelArray = [string, string, string][];
+export type Target = {
+  label: string;
+  description: string;
+};
