+++ src/utils.ts
@@ -11,8 +11,6 @@ import { cli } from "./cli";
 export {
   getDir,
   installPretext,
-  ptxExec,
-  getTargets,
   setSchema,
   setSpellCheckConfig,
   updateStatusBarItem,
@@ -22,15 +20,6 @@ export {
 };
 
 async function experiment() {
-  console.log("Current pythonPath:", cli._pythonPath);
-  console.log("cli.pythonPath(): ", cli.pythonPath());
-  console.log("cli._pythonPath: ", cli._pythonPath);
-  console.log("cli.pythonPath('test'): ", cli.pythonPath('test'));
-  console.log("cli._pythonPath: ", cli._pythonPath);
-  cli._pythonPath = "hello";
-  console.log("cli._pythonPath: ", cli._pythonPath);
-  console.log("cli.pythonPath(): ", cli.pythonPath());
-  console.log("cli._pythonPath: ", cli.pythonPath(null));
   return;
 }
 
@@ -70,8 +59,8 @@ async function installPretext(progress: vscode.Progress<{}>) {
   // Here we will attempt to pip install pretext, upgraded to the most recent version.  This will happen if pretext is not found, or if a user requests it through a command.
 
   // first check for python and pip:
-  let pythonExec = getPythonExec();
-  if (pythonExec === undefined) {
+  let pythonExec = cli.pythonPath();
+  if (!pythonExec) {
     vscode.window.showErrorMessage(
       "Unable to install PreTeXt without python.  Please install python and try again."
     );
@@ -117,165 +106,8 @@ async function installPretext(progress: vscode.Progress<{}>) {
 /**
  * Get python command, either from settings or by verifying python3 or python is on PATH
  */
-function getPythonExec() {
-  let pythonExec = vscode.workspace
-    .getConfiguration("pretext-tools")
-    .get("pythonPath");
-  console.log("Python path from settings: ", pythonExec);
-  if (pythonExec === "") {
-    for (let command of ["python3", "python"]) {
-      try {
-        let pythonVersion = execSync(command + " --version").toString();
-        console.log("Python version result: ", pythonVersion);
-        if (pythonVersion.toLowerCase().includes("python 2")) {
-          throw new Error(command + " is python 2");
-        }
-        pythonExec = command;
-        break;
-      } catch (err) {
-        console.log("Error: ", err);
-      }
-    }
-  } else {
-    try {
-      let pythonVersion = execSync(pythonExec + " --version").toString();
-      console.log("Python version result: ", pythonVersion);
-      if (pythonVersion.toLowerCase().includes("python 2")) {
-        throw new Error(pythonExec + " is python 2");
-      }
-    } catch (err) {
-      console.log("Error: ", err);
-      vscode.window.showErrorMessage(
-        "The path to python provided in settings does not appear to be a valid python 3 executable.  Please provide a valid path to python."
-      );
-      pythonExec = "";
-    }
-  }
-  if (pythonExec === "") {
-    vscode.window.showErrorMessage(
-      "You do not appear to have python installed.  Please download and install python (and make sure it is added to your PATH)."
-    );
-    return undefined;
-  }
-  return pythonExec;
-}
 
-function getPtxExec() {
-  let pythonExec = getPythonExec();
-  if (pythonExec === undefined) {
-    vscode.window.showErrorMessage(
-      "Unable to run PreTeXt without python.  Please install python and try again."
-    );
-  } else {
-    console.log("Using python at ", pythonExec);
-    let ptxCommand = pythonExec + " -m pretext";
-    try {
-      let ptxVersion = execSync(ptxCommand + " --version").toString();
-      console.log("Using PreTeXt version", ptxVersion);
-      return ptxCommand;
-    } catch (err) {
-      console.log(ptxCommand + " not found");
-    }
-  }
-  // If the above did not work, then either pretext has been installed with pipx, is not yet installed, or python is not available in the usual place, so we try a last ditch effort to find it.
-  let ptxExec = "";
-  for (let command of ["pretext", "python -m pretext", "python3 -m pretext"]) {
-    try {
-      let ptxVersion = execSync(command + " --version").toString();
-      console.log("Using PreTeXt version", ptxVersion);
-      return command;
-    } catch (err) {
-      console.log(command + " not found");
-    }
-  }
-  if (
-    ptxExec === "" &&
-    vscode.workspace.getConfiguration("pretext-tools").get("installPretext")
-  ) {
-    vscode.window
-      .showWarningMessage(
-        "It doesn't look like you have pretext installed.  Would you like to try to install it now?",
-        "Yes",
-        "No",
-        "No (stop asking)"
-      )
-      .then((option) => {
-        if (option === "Yes") {
-          try {
-            vscode.commands.executeCommand("pretext-tools.updatePTX");
-            console.log("Finished attempting to install PreTeXt");
-            return getPtxExec();
-          } catch (err) {
-            console.log("Unable to install PreTeXt.  Error: ", err);
-          }
-        } else if (option === "No (stop asking)") {
-          vscode.workspace
-            .getConfiguration("pretext-tools")
-            .update("installPretext", false);
-        }
-      });
-  }
-  return ptxExec;
-}
-
-function getPtxVersion() {
-  let ptxVersion = "version unknown";
-  try {
-    ptxVersion = execSync(ptxExec + " --version")
-      .toString()
-      .trim();
-    if (ptxVersion.includes("\n")) {
-      ptxVersion = ptxVersion.split("\n")[-1];
-    }
-    console.log("Using PreTeXt version", ptxVersion);
-  } catch (err) {
-    console.log("Error: ", err);
-  }
-  return ptxVersion;
-}
-
-const ptxExec = getPtxExec();
-const ptxVersion = getPtxVersion();
-
-/**
- * Get list of targets from project.ptx file
- * @returns Array of target names
- */
-function getTargets() {
-  // execSync returns stdout from executing the command.  We then convert to a string, split on new lines, and remove any blanks, to create an array of the target names.
-  // See also https://stackoverflow.com/questions/41001360/saving-the-output-of-a-child-process-in-a-variable-in-the-parent-in-nodejs
-  let filePath = getDir();
-  try {
-    let targets = execSync(ptxExec + " --targets", { cwd: filePath })
-      .toString()
-      .split(/\r?\n/)
-      .filter(Boolean);
-    // Set up dictionary for quickselect:
-    if (targets.length > 0) {
-      let targetSelection = [];
-      for (let target of targets) {
-        // exclude lines that start with "Generated" as these are not targets
-        if (!target.includes("Generated")) {
-          targetSelection.push({
-            label: target,
-            description: "Build source as " + target,
-          });
-        }
-      }
-      return targetSelection;
-    } else {
-      return [
-        {
-          label: "No PreTeXt project found.",
-          description: "Change to directory containing a project.ptx file.",
-        },
-      ];
-    }
-  } catch (err) {
-    console.log("getTargets() Error: \n", err);
-    return [];
-  }
-}
+//const ptxVersion = getPtxVersion();
 
 function setSpellCheckConfig() {
   const cSpellConfig = vscode.workspace.getConfiguration("cSpell");
@@ -332,20 +164,13 @@ function setSchema() {
   let schemaPath: string | undefined = vscode.workspace
     .getConfiguration("pretext-tools")
     .get("schema.customPath");
+
   if (schemaPath === "") {
-    const userHomeDir: string = homedir();
+    const extensionPath = path.resolve(__dirname);
+    let schemaDir = path.join(extensionPath, "assets", "schema");
     const schemaConfig = vscode.workspace
       .getConfiguration("pretext-tools")
       .get("schema.versionName");
-    // set schema folder based on ptxVersion number:
-    //  - < 2.5, use userHomeDir/.ptx/schema/
-    //  - >= 2.5, use userHomeDir/.ptx/{ptxVersion}/core/schema/
-    let schemaDir: string;
-    if (parseFloat(ptxVersion) < 2.5) {
-      schemaDir = path.join(userHomeDir, ".ptx", "schema");
-    } else {
-      schemaDir = path.join(userHomeDir, ".ptx", ptxVersion, "core", "schema");
-    }
     switch (schemaConfig) {
       case "Stable":
         schemaPath = path.join(schemaDir, "pretext.rng");
@@ -357,11 +182,11 @@ function setSchema() {
         console.log(
           "Selected custom schema, but no path provided.  Setting to default."
         );
-        schemaPath =
-          "https://raw.githubusercontent.com/PreTeXtBook/pretext/master/schema/pretext.rng";
+        schemaPath = path.join(schemaDir, "pretext.rng");
         break;
     }
   }
+  console.log("Schema set to: ", schemaPath);
   const configuration = vscode.workspace.getConfiguration("xml");
   let schemas: any = configuration.get("fileAssociations");
   for (let dicts of schemas) {
@@ -383,7 +208,7 @@ function updateStatusBarItem(
   ptxSBItem.show();
   if (state === "ready" || state === undefined) {
     ptxSBItem.text = `$(debug-run) PreTeXt`;
-    ptxSBItem.tooltip = `Run PreTeXt command (version ${ptxVersion})`;
+    ptxSBItem.tooltip = `Run PreTeXt command`;
     ptxSBItem.command = `pretext-tools.selectPretextCommand`;
   } else if (state === "running") {
     ptxSBItem.text = `$(loading~spin) PreTeXt`;
